refactor: rename broken_link_checker by blc a short easier to use
