+++ tests/checker_test.sh
@@ -33,7 +33,7 @@ start_server() {
 
 # We start the test
 start_test() {
-    report=$($PYTHON -m broken_link_checker http://$HOST:$PORT -D -d 0 $BLC_FLAGS)
+    report=$($PYTHON -m blc http://$HOST:$PORT -D -d 0 $BLC_FLAGS)
     nb_broken_link_got=$(expr $(echo "$report" | grep -c .) - 2)
     if [ ! $nb_broken_link_got -eq $NB_BROKEN_LINK_EXPECTED ]; then
         echo "$NB_BROKEN_LINK_EXPECTED broken links expected, but $nb_broken_link_got got"
